requirements.txt
psycopg2-binary==2.9.10
ollama==0.5.1
pypdf==5.8.0
